Refactor MonoGameGumInCode sample with per-screen classes and nav strip
Game1 previously held all layout-construction logic inline and required
editing/recompiling to switch between demos. Extract each layout into its
own FrameworkElement-derived screen under Screens/ (matching the existing
FormsScreen / StandardsScreen pattern), add a NineSliceScreen that walks
the runtime's full property surface (default stretching, TextureAddress.
Custom, color tinting, IsTilingMiddleSections), and add an in-app nav
strip of Forms Buttons that swaps the active screen at runtime.

Closes #2736
